refactor(NameTest): 이름 관련 테스트를 자동차에서 분리
